+++ ppaya_lambda_utils/exceptions.py
@@ -1,0 +1,3 @@
+
+class InvokeLambdaFunctionException(Exception):
+    """ Raised when the invocation of a lamda function fails """