+++ tests/test_boto_utils.py
@@ -1,8 +1,13 @@
 from __future__ import annotations
+import io
+from ppaya_lambda_utils.exceptions import InvokeLambdaFunctionException
 from typing import List, TYPE_CHECKING
+from unittest.mock import Mock
+
+import pytest
 
 from ppaya_lambda_utils.boto_utils import (
-    send_to_sqs, boto_clients, publish_to_sns)
+    invoke_lambda_function, send_to_sqs, boto_clients, publish_to_sns)
 
 if TYPE_CHECKING:
     from mypy_boto3_sqs.type_defs import SendMessageBatchRequestEntryTypeDef
@@ -56,3 +61,33 @@ def test_publish_to_sns_without_attributes(sns_topic, sns_subscription) -> None:
 
     message = sns_subscription.receive_messages(MaxNumberOfMessages=1)[0]
     assert message == message
+
+
+def test_invoke_lambda_function_with_event_invocation_type(mocker):
+    mock_lambda_client = Mock()
+    mock_lambda_client().invoke.return_value = {'StatusCode': 202}
+    mocker.patch.object(boto_clients, 'get_client', mock_lambda_client)
+
+    invoke_lambda_function('test_func', {'msg': 'blah'})
+
+
+def test_invoke_lambda_function_with_error(mocker):
+    mock_lambda_client = Mock()
+    mock_lambda_client().invoke.return_value = {'StatusCode': 400}
+    mocker.patch.object(boto_clients, 'get_client', mock_lambda_client)
+
+    with pytest.raises(InvokeLambdaFunctionException):
+        invoke_lambda_function('test_func', {'msg': 'blah'})
+
+
+def test_invoke_lambda_function_with_request_response_invocation_type(mocker):
+    mock_lambda_client = Mock()
+    mock_lambda_client().invoke.return_value = {
+        'StatusCode': 200,
+        'Payload': io.StringIO('{"x": 1}'),
+    }
+    mocker.patch.object(boto_clients, 'get_client', mock_lambda_client)
+
+    resp = invoke_lambda_function('test_func', {'msg': 'blah'}, 'RequestResponse')
+
+    assert resp == {'x': 1}